Refactor gender options: move to consts/index.ts and remove unused index.tsx
