+++ consts/index.ts
@@ -1,3 +1,5 @@
+import { SelectionOption } from "@/types/selection-option.types";
+
 export const BUTTON_CLASS_NAMES = {
   Primary: "bg-primary text-white",
   Secondary: "bg-secondary text-white",
@@ -11,3 +13,18 @@ export const BUTTON_CLASS_NAMES = {
   Warning: "bg-warning text-white",
   Info: "bg-info text-white",
 };
+
+export const GENDER_OPTIONS: SelectionOption[] = [
+  {
+    label: "Man",
+    value: "male",
+  },
+  {
+    label: "Vrouw",
+    value: "female",
+  },
+  {
+    label: "Niet gespecificeerd",
+    value: "not_specified",
+  },
+];
