--- consts/index.tsx
@@ -1,16 +1,0 @@
-import { SelectionOption } from "@/types/selection-option.types";
-
-export const GENDER_OPTIONS: SelectionOption[] = [
-    {
-        label: "Man",
-        value: "male",
-    },
-    {
-        label: "Vrouw",
-        value: "female",
-    },
-    {
-        label: "Niet gespecificeerd",
-        value: "not_specified",
-    },
-];
